+++ test/fixture/mock-generator/variadics/expected.php
@@ -101,6 +101,38 @@ implements \Eloquent\Phony\Mock\Mock
         return $result;
     }
 
+    public function methodD(
+        $a0,
+        $a1,
+        ?\stdClass ...$a2
+    ) {
+        $argumentCount = \func_num_args();
+        $arguments = [];
+
+        if ($argumentCount > 0) {
+            $arguments[] = $a0;
+        }
+        if ($argumentCount > 1) {
+            $arguments[] = $a1;
+        }
+
+        for ($i = 2; $i < $argumentCount; ++$i) {
+            $arguments[] = $a2[$i - 2];
+        }
+
+        if (!$this->_handle) {
+            $result = null;
+
+            return $result;
+        }
+
+        $result = $this->_handle->spy(__FUNCTION__)->invokeWith(
+            new \Eloquent\Phony\Call\Arguments($arguments)
+        );
+
+        return $result;
+    }
+
     private static $_uncallableMethods = [];
     private static $_traitMethods = [];
     private static $_customMethods = [];