+++ test/src/Test/TestInterfaceWithVariadicParameterWithNullableType.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Eloquent\Phony\Test;
+
+use stdClass;
+
+interface TestInterfaceWithVariadicParameterWithNullableType
+{
+    public function method(?stdClass ...$arguments);
+}